requirements.txt
prowlarr-py==1.0.1
pyarr==5.2.0
PyYAML==6.0